Resizing of the SVG in the preload
One of the main qualities of SVG images, is its scalability, not being able to use this, is to reduce it to be a bit more image. An SVG file has an XML format, so it can be modified with the DOM.
In order to collaborate, I have put the option to modify the parameters of the SVG and in this way make it adaptable.
The modification gives the option of being able to choose the size in the preload to which you want to have the image before rasterizing it; And so avoid the subsequent scaling that will not have the same quality.
You can see the example here: 
https://www.youtube.com/watch?v=Kd7pNjY-FLk
tested in version 3.11 It does not interfere with the original code.